requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-ollama>=0.1.0
langchain-community>=0.3.0
mem0ai>=1.0.5
faiss-cpu>=1.7.0
pypdf>=4.0.0
streamlit>=1.32.0
sentence-transformers>=2.2.0
chromadb>=0.4.0
tiktoken>=0.5.0
langgraph-checkpoint-sqlite
langchain-text-splitters
pymupdf 